+++ packages/logo/src/Logo.types.ts
@@ -4,9 +4,13 @@ export type LogoProps = {
    */
   fontSize?: number;
   /**
-   * The color of the logo primary secondary and tertiary
+   * @deprecated, use 'usage' instead. The color of the logo primary secondary and tertiary
    */
   color?: string;
+  /**
+   * Selects the correct logo color way based on the type of color it is used on
+   */
+  usage?: 'dark_bg' | 'light_bg' | 'mint_bg';
   /**
    * if true hide non-interactive content from the accessibility API
    */